Date utils separated from EventList
